+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/AnalyzeInvoicesFromUrlAsync.java
@@ -119,7 +119,7 @@ public class AnalyzeInvoicesFromUrlAsync {
                 DocumentField invoiceTotalField = invoiceFields.get("InvoiceTotal");
                 if (customerAddressRecipientField != null) {
                     if (DocumentFieldType.DOUBLE == invoiceTotalField.getType()) {
-                        Float invoiceTotal = invoiceTotalField.getValueAsDouble();
+                        Double invoiceTotal = invoiceTotalField.getValueAsDouble();
                         System.out.printf("Invoice Total: %.2f, confidence: %.2f%n",
                             invoiceTotal, invoiceTotalField.getConfidence());
                     }
@@ -145,21 +145,21 @@ public class AnalyzeInvoicesFromUrlAsync {
                                 }
                                 if ("Quantity".equals(key)) {
                                     if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                        Float quantity = documentField.getValueAsDouble();
+                                        Double quantity = documentField.getValueAsDouble();
                                         System.out.printf("Quantity: %f, confidence: %.2f%n",
                                             quantity, documentField.getConfidence());
                                     }
                                 }
                                 if ("UnitPrice".equals(key)) {
                                     if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                        Float unitPrice = documentField.getValueAsDouble();
+                                        Double unitPrice = documentField.getValueAsDouble();
                                         System.out.printf("Unit Price: %f, confidence: %.2f%n",
                                             unitPrice, documentField.getConfidence());
                                     }
                                 }
                                 if ("ProductCode".equals(key)) {
                                     if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                        Float productCode = documentField.getValueAsDouble();
+                                        Double productCode = documentField.getValueAsDouble();
                                         System.out.printf("Product Code: %f, confidence: %.2f%n",
                                             productCode, documentField.getConfidence());
                                     }
