+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/AnalyzeReceipts.java
@@ -108,21 +108,21 @@ public class AnalyzeReceipts {
                             }
                             if ("Quantity".equals(key)) {
                                 if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                    Float quantity = documentField.getValueAsDouble();
+                                    Double quantity = documentField.getValueAsDouble();
                                     System.out.printf("Quantity: %f, confidence: %.2f%n",
                                         quantity, documentField.getConfidence());
                                 }
                             }
                             if ("Price".equals(key)) {
                                 if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                    Float price = documentField.getValueAsDouble();
+                                    Double price = documentField.getValueAsDouble();
                                     System.out.printf("Price: %f, confidence: %.2f%n",
                                         price, documentField.getConfidence());
                                 }
                             }
                             if ("TotalPrice".equals(key)) {
                                 if (DocumentFieldType.DOUBLE == documentField.getType()) {
-                                    Float totalPrice = documentField.getValueAsDouble();
+                                    Double totalPrice = documentField.getValueAsDouble();
                                     System.out.printf("Total Price: %f, confidence: %.2f%n",
                                         totalPrice, documentField.getConfidence());
                                 }
